requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
streamlit==1.28.0
plotly==5.17.0
Pillow==10.0.0
